Remove redundant builtins
